Fix `thread_state64` on `aarch64-apple-darwin`
